Add files via upload
These were the files developed by myself. There are a few files inspired and taken from aoxolotl's Head Detector Repository that I, later on, tweaked a bit with the detection settings and changed a few things related to the code itself. This code is not permanent and will make further improvements in order to, hopefully, track a large number of people without much error while exporting all detection values to a single file or to a database.